DOC: require Sphinx >= 1.5.1
Older versions of Sphinx are incompatible with newer versions of
docutils.
It is still possible to install nbsphinx on old systems, as long as docutils is
also old.

The Travis-CI settings were adapted to install old docutils when testing old
Sphinx.

The ban of Sphinx 1.5.0 was removed from setup.py.
This wasn't quite correct anyway, since all versions below 1.5.1 are affected
by the docutils incompatibility.