Make implicit imports explicit
Closes #825

Implicit imports (imports that rely on the fact
that importing a module imports another module
such as a submodule) makes the code vulnerable
to breakage due to refactors in the dependent
code. Such as the changes in Sphinx 8.2.0 that
caused #825

This replaces all implicit imports that I
could find by explicit imports.